backend/requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
bcrypt
streamlit
PyJWT
streamlit-aggrid
XlsxWriter
openpyxl